api/requirements.txt
amqp==2.6.1
anyio==3.6.1
billiard==3.6.4.0
celery==4.4.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.82.0
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
kombu==4.6.11
loguru==0.6.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pydantic==1.10.2
pyhumps==3.7.3
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
redis==3.5.3
requests==2.28.1
serpy==0.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
vine==1.3.0
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3

pyTelegramBotAPI==3.7.4
structlog
